move configuration files tab to configuraon page show configuraton
