+++ Source/SelfService/Web/microservice/base/configuration.tsx
@@ -1,9 +1,9 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import { Box, Divider } from '@mui/material';
+import { Box, Divider, List, Typography } from '@mui/material';
 import React, {useEffect, useState} from 'react';
-import { IngressURLWithCustomerTenantID, SimpleIngressPath } from '../../api/api';
+import { getConfigFilesNamesList, IngressURLWithCustomerTenantID, SimpleIngressPath, updateConfigFiles } from '../../api/api';
 import { MicroserviceSimple } from '../../api/index';
 
 
@@ -21,23 +21,70 @@ export type ConfigurationProps = {
     ingressPaths: SimpleIngressPath[]
     ms: MicroserviceSimple
     onClick: () => void
-// add onclick
-
 };
 
 
-
-const MAX_CONFIGMAP_ENTRY_SIZE = 3145728;
-
-
 const styles = {
+    base: {
+        '& .MuiTypography-body1': {
+            color: 'black',
+            align: 'center'
+        }
+    },
     divider: {
         backgroundColor: '#3B3D48'
     }
 };
 
+const MAX_CONFIGMAP_ENTRY_SIZE = 3145728;
+
 
 export const Configuration: React.FunctionComponent<ConfigurationProps> = (props) => {
+    const [filesNamesList, setFilesNamesList] = useState<string[]>([]);
+    async function onFileSelectorSubmit(event) {
+        event.preventDefault();
+
+        const fData = new FormData(event.target);
+
+        await updateConfigFiles(props.applicationId, props.environment, props.microserviceId, fData);
+        fetchConfigFilesNamesList();
+    }
+
+    const fileSelector = document?.getElementById('config-file-selector-form');
+
+    const attachFormSubmitEvent = () => {
+        fileSelector?.addEventListener('submit', onFileSelectorSubmit);
+    }
+    attachFormSubmitEvent();
+
+
+    fileSelector?.addEventListener('change', (event: any) => {
+        const fileList = event.target.files;
+        if (fileList[0].size > MAX_CONFIGMAP_ENTRY_SIZE) {
+            alert('file cannot be larger than 3145728 bytes. Please select another file');
+        }
+    });
+
+
+
+
+    useEffect(() => {
+        fetchConfigFilesNamesList()
+            .catch(console.error);;
+
+    }, []);
+
+    const fetchConfigFilesNamesList = async () => {
+        const result = await getConfigFilesNamesList(props.applicationId, props.environment, props.microserviceId);
+
+        if (!result.data) {
+            window.alert(`Could not fetch config files`);
+        }
+
+        setFilesNamesList(result.data);
+    };
+
+
 
     return(
         <>
@@ -60,7 +107,44 @@ export const Configuration: React.FunctionComponent<ConfigurationProps> = (props
                             applicationId={props.applicationId}
                         />
             </Box>
-
+            <Box ml={2}>
+                <Typography
+                    variant="h4"
+                    component="h4"
+                    style={{
+                        paddingBottom: '5px',
+                    }}>Microservice configuration files</Typography>
+                <Typography
+                    variant="body2"
+                    component="p"
+                    style={{
+                        paddingBottom: '50px',
+                    }}>/app/data</Typography>
+                {/* <List items={filesNamesList} onRenderCell={onRenderCell} style={{ backgroundColor: 'white', width: '30%' }} /> */}
+                <Divider style={{ backgroundColor: '#3B3D48', marginTop: '40px', marginBottom: '20px' }} />
+                <Typography
+                    variant="h4"
+                    component="h4"
+                    style={{
+                        paddingBottom: '5px',
+                    }}>Add new configuration file</Typography>
+                <Typography
+                    variant="body2"
+                    component="p"
+                    style={{
+                        paddingBottom: '5px',
+                    }}>max file size: {MAX_CONFIGMAP_ENTRY_SIZE} bytes / 3.15mb</Typography>
+                <Typography
+                    variant="body2"
+                    component="p"
+                    style={{
+                        paddingBottom: '50px',
+                    }}>submitting files with the same name will cause the original file to be replaced</Typography>
+                <form method="put" id="config-file-selector-form">
+                    <input type="file" id="file-selector" name='file' />
+                    <input type="submit" value="Submit" />
+                </form>
+            </Box>
         </>
     )
 }
